Added logic for custom react agent and standard_react_agent
