+++ docker-compose.yaml
@@ -5,7 +5,7 @@ services:
     image: redis:6-alpine
     container_name: langfold-redis
     pull_policy: if_not_present
-    command: redis-server --requirepass ${REDIS_PASSWORD:-password} --port ${REDIS_PORT:-6380}
+    command: redis-server --requirepass ${REDIS_PASSWORD:-password} --port ${REDIS_PORT:-6380} --loglevel debug
     ports:
       - "${REDIS_PORT:-6380}:6380"
     healthcheck:
