+++ src/nodes/call_model.py
@@ -1,6 +1,6 @@
 from langchain_core.runnables import RunnableConfig
 from src.models.state import AgentState
-from src.services.llm import MODEL
+from src.services.agents.model_provider import TOOL_ENABLED_MODEL
 
 
 async def call_model(
@@ -23,7 +23,7 @@ async def call_model(
               expected to be a single message generated by the model based on the input messages.
     """
     # Invoke the language model asynchronously with the messages from the state and the provided configuration
-    response = await MODEL.ainvoke(state["messages"], config)
+    response = await TOOL_ENABLED_MODEL.ainvoke(state["messages"], config)
 
     # Return the response in a structured format, wrapping it in a list under the 'messages' key
     return {"messages": [response]}
