+++ src/services/agents/__init__.py
@@ -1,0 +1,4 @@
+from src.services.agents.agent_factory import run_agent
+from src.services.agents.model_provider import MODEL, TOOL_ENABLED_MODEL
+
+__all__ = ["run_agent", "MODEL", "TOOL_ENABLED_MODEL"]
